rename views for notes
