+++ web-client/src/views/CaseDetail/DeleteCaseNoteConfirmModal.jsx
